requirements.txt
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
binaryornot==0.4.4
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
cookiecutter==2.3.0
crispy-tailwind==0.5.0
cryptography==41.0.4
cssselect2==0.7.0
Django==4.2
django-browser-reload==1.8.0
django-crispy-forms==1.13.0
django-easy-pdf==0.1.1
django-tailwind @ git+https://github.com/timonweb/django-tailwind.git@c218f008a9c99a2dd3ebc6300a8314bc550fc78b
django-weasyprint==2.2.1
fonttools==4.42.1
html5lib==1.1
idna==3.4
Jinja2==3.1.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
oscrypto==1.3.0
Pillow==10.0.0
pycparser==2.21
pydyf==0.7.0
Pygments==2.16.1
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==3.16.1
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.8.0
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
weasyprint==59.0
webencodings==0.5.1
xhtml2pdf==0.2.11
zopfli==0.2.3
